pyproject.toml
[tool.poetry]
name = "cv-pruning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include = "src"},
    {include = "configs"},
]

[tool.poetry.dependencies]
python = "~3.9"
matplotlib = "*"
numpy = "*"
Pillow = "*"
PyYAML = "*"
tqdm = "*"
torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp39-cp39-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp39-cp39-linux_x86_64.whl"}
pandas = "*"
ipykernel = "^6.29.2"
opencv-python = "^4.9.0.80"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
seaborn = "^0.13.2"
scipy = "^1.12.0"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"